+++ src/app/spa-auth/[locale]/[[...path]]/route.ts
@@ -1,11 +1,10 @@
 import { getServerFeatureFlagsValue } from '@/config/featureFlags';
-import { analyticsEnv } from '@/envs/analytics';
 import { appEnv } from '@/envs/app';
 import { authEnv } from '@/envs/auth';
 import { type Locales, normalizeLocale } from '@/locales/resources';
 import { getServerAuthConfig } from '@/server/globalConfig/getServerAuthConfig';
-import { serializeForHtml } from '@/server/utils/serializeForHtml';
-import { type AnalyticsConfig, type AuthSPAServerConfig } from '@/types/spaServerConfig';
+import { buildAnalyticsConfig, fetchViteDevTemplate, renderSpaHtml } from '@/server/spaHtml';
+import { type AuthSPAServerConfig } from '@/types/spaServerConfig';
 
 import { buildSeoMeta } from './seoMeta';
 
@@ -16,123 +15,13 @@ export function generateStaticParams() {
 }
 
 const isDev = process.env.NODE_ENV === 'development';
-const VITE_DEV_ORIGIN = 'http://localhost:9876';
-
-async function rewriteViteAssetUrls(html: string): Promise<string> {
-  const { parseHTML } = await import('linkedom');
-  const { document } = parseHTML(html);
-
-  document.querySelectorAll('script[src]').forEach((el: Element) => {
-    const src = el.getAttribute('src');
-    if (src && src.startsWith('/')) {
-      el.setAttribute('src', `${VITE_DEV_ORIGIN}${src}`);
-    }
-  });
-
-  document.querySelectorAll('link[href]').forEach((el: Element) => {
-    const href = el.getAttribute('href');
-    if (href && href.startsWith('/')) {
-      el.setAttribute('href', `${VITE_DEV_ORIGIN}${href}`);
-    }
-  });
-
-  document.querySelectorAll('script[type="module"]:not([src])').forEach((el: Element) => {
-    const text = el.textContent || '';
-    if (text.includes('/@')) {
-      el.textContent = text.replaceAll(
-        /from\s+["'](\/[@\w].*?)["']/g,
-        (_match: string, p: string) => `from "${VITE_DEV_ORIGIN}${p}"`,
-      );
-    }
-  });
-
-  const workerPatch = document.createElement('script');
-  workerPatch.textContent = `(function(){
-var O=globalThis.Worker;
-globalThis.Worker=function(u,o){
-var h=typeof u==='string'?u:u instanceof URL?u.href:'';
-if(h.startsWith('${VITE_DEV_ORIGIN}')){
-var b=new Blob(['import "'+h+'";'],{type:'application/javascript'});
-return new O(URL.createObjectURL(b),Object.assign({},o,{type:'module'}));
-}return new O(u,o)};
-globalThis.Worker.prototype=O.prototype;
-})();`;
-  const head = document.querySelector('head');
-  if (head?.firstChild) {
-    head.insertBefore(workerPatch, head.firstChild);
-  }
-
-  return document.toString();
-}
 
 async function getTemplate(): Promise<string> {
-  if (isDev) {
-    const res = await fetch(`${VITE_DEV_ORIGIN}/index.auth.html`);
-    const html = await res.text();
-    return rewriteViteAssetUrls(html);
-  }
+  if (isDev) return fetchViteDevTemplate('/index.auth.html');
 
   const { authHtmlTemplate } = await import('../../authHtmlTemplate');
 
   return authHtmlTemplate;
-}
-
-function buildAnalyticsConfig(): AnalyticsConfig {
-  const config: AnalyticsConfig = {};
-
-  if (analyticsEnv.ENABLE_GOOGLE_ANALYTICS && analyticsEnv.GOOGLE_ANALYTICS_MEASUREMENT_ID) {
-    config.google = { measurementId: analyticsEnv.GOOGLE_ANALYTICS_MEASUREMENT_ID };
-  }
-
-  if (analyticsEnv.ENABLED_PLAUSIBLE_ANALYTICS && analyticsEnv.PLAUSIBLE_DOMAIN) {
-    config.plausible = {
-      domain: analyticsEnv.PLAUSIBLE_DOMAIN,
-      scriptBaseUrl: analyticsEnv.PLAUSIBLE_SCRIPT_BASE_URL,
-    };
-  }
-
-  if (analyticsEnv.ENABLED_UMAMI_ANALYTICS && analyticsEnv.UMAMI_WEBSITE_ID) {
-    config.umami = {
-      scriptUrl: analyticsEnv.UMAMI_SCRIPT_URL,
-      websiteId: analyticsEnv.UMAMI_WEBSITE_ID,
-    };
-  }
-
-  if (analyticsEnv.ENABLED_CLARITY_ANALYTICS && analyticsEnv.CLARITY_PROJECT_ID) {
-    config.clarity = { projectId: analyticsEnv.CLARITY_PROJECT_ID };
-  }
-
-  if (analyticsEnv.ENABLED_POSTHOG_ANALYTICS && analyticsEnv.POSTHOG_KEY) {
-    config.posthog = {
-      debug: analyticsEnv.DEBUG_POSTHOG_ANALYTICS,
-      host: analyticsEnv.POSTHOG_HOST,
-      key: analyticsEnv.POSTHOG_KEY,
-    };
-  }
-
-  if (analyticsEnv.ENABLED_X_ADS && analyticsEnv.X_ADS_PIXEL_ID) {
-    config.xAds = {
-      eventIds: {
-        login_or_signup_clicked: analyticsEnv.X_ADS_LOGIN_OR_SIGNUP_CLICKED_EVENT_ID,
-        main_page_view: analyticsEnv.X_ADS_MAIN_PAGE_VIEW_EVENT_ID,
-      },
-      pixelId: analyticsEnv.X_ADS_PIXEL_ID,
-      purchaseEventId: analyticsEnv.X_ADS_PURCHASE_EVENT_ID,
-    };
-  }
-
-  if (analyticsEnv.REACT_SCAN_MONITOR_API_KEY) {
-    config.reactScan = { apiKey: analyticsEnv.REACT_SCAN_MONITOR_API_KEY };
-  }
-
-  if (analyticsEnv.ENABLE_VERCEL_ANALYTICS) {
-    config.vercel = {
-      debug: analyticsEnv.DEBUG_VERCEL_ANALYTICS,
-      enabled: true,
-    };
-  }
-
-  return config;
 }
 
 export async function GET(
@@ -150,22 +39,9 @@ export async function GET(
     globalCDN: appEnv.CDN_USE_GLOBAL,
   };
 
-  let html = await getTemplate();
-
-  html = html.replace(
-    /window\.__SERVER_CONFIG__\s*=\s*undefined;\s*\/\*\s*SERVER_CONFIG\s*\*\//,
-    `window.__SERVER_CONFIG__ = ${serializeForHtml(authConfig)};`,
-  );
-
+  const template = await getTemplate();
   const pathname = `/${(path ?? []).join('/')}`;
   const seoMeta = await buildSeoMeta(locale, pathname);
-  html = html.replace('<!--SEO_META-->', seoMeta);
-  html = html.replace('<!--ANALYTICS_SCRIPTS-->', '');
 
-  return new Response(html, {
-    headers: {
-      'Cache-Control': 'no-cache',
-      'content-type': 'text/html; charset=utf-8',
-    },
-  });
+  return renderSpaHtml(template, { seoMeta, serverConfig: authConfig });
 }
